split ast into multiple files
