+++ pkg/ast/ast_schema_definition.go
@@ -1,0 +1,24 @@
+package ast
+
+import "github.com/jensneuse/graphql-go-tools/pkg/lexer/position"
+
+type SchemaDefinition struct {
+	SchemaLiteral                position.Position // schema
+	HasDirectives                bool
+	Directives                   DirectiveList                   // optional, e.g. @foo
+	RootOperationTypeDefinitions RootOperationTypeDefinitionList // e.g. query: Query, mutation: Mutation, subscription: Subscription
+}
+
+func (s *SchemaDefinition) AddRootOperationTypeDefinitionRefs(refs ...int) {
+	s.RootOperationTypeDefinitions.Refs = append(s.RootOperationTypeDefinitions.Refs, refs...)
+}
+
+func (d *Document) AddSchemaDefinition(schemaDefinition SchemaDefinition) (ref int) {
+	d.SchemaDefinitions = append(d.SchemaDefinitions, schemaDefinition)
+	return len(d.SchemaDefinitions) - 1
+}
+
+type SchemaExtension struct {
+	ExtendLiteral position.Position
+	SchemaDefinition
+}
