+++ pkg/ast/ast_value.go
@@ -1,0 +1,157 @@
+package ast
+
+import (
+	"bytes"
+	"fmt"
+	"io"
+
+	"github.com/jensneuse/graphql-go-tools/pkg/lexer/literal"
+)
+
+type ValueKind int
+
+const (
+	ValueKindUnknown ValueKind = iota
+	ValueKindString
+	ValueKindBoolean
+	ValueKindInteger
+	ValueKindFloat
+	ValueKindVariable
+	ValueKindNull
+	ValueKindList
+	ValueKindObject
+	ValueKindEnum
+)
+
+type Value struct {
+	Kind ValueKind // e.g. 100 or "Bar"
+	Ref  int
+}
+
+func (d *Document) ValueContentBytes(value Value) ByteSlice {
+	switch value.Kind {
+	case ValueKindEnum:
+		return d.EnumValueNameBytes(value.Ref)
+	case ValueKindString:
+		d.StringValueContentBytes(value.Ref)
+	case ValueKindInteger:
+		return d.IntValueRaw(value.Ref)
+	case ValueKindFloat:
+		return d.FloatValueRaw(value.Ref)
+	}
+	panic(fmt.Errorf("ValueContentBytes not implemented for ValueKind: %s", value.Kind))
+}
+
+// nolint
+func (d *Document) PrintValue(value Value, w io.Writer) (err error) {
+	switch value.Kind {
+	case ValueKindBoolean:
+		if d.BooleanValues[value.Ref] {
+			_, err = w.Write(literal.TRUE)
+		} else {
+			_, err = w.Write(literal.FALSE)
+		}
+	case ValueKindString:
+		_, err = w.Write(literal.QUOTE)
+		_, err = w.Write(d.Input.ByteSlice(d.StringValues[value.Ref].Content))
+		_, err = w.Write(literal.QUOTE)
+	case ValueKindInteger:
+		if d.IntValues[value.Ref].Negative {
+			_, err = w.Write(literal.SUB)
+		}
+		_, err = w.Write(d.Input.ByteSlice(d.IntValues[value.Ref].Raw))
+	case ValueKindFloat:
+		if d.FloatValues[value.Ref].Negative {
+			_, err = w.Write(literal.SUB)
+		}
+		_, err = w.Write(d.Input.ByteSlice(d.FloatValues[value.Ref].Raw))
+	case ValueKindVariable:
+		_, err = w.Write(literal.DOLLAR)
+		_, err = w.Write(d.Input.ByteSlice(d.VariableValues[value.Ref].Name))
+	case ValueKindNull:
+		_, err = w.Write(literal.NULL)
+	case ValueKindList:
+		_, err = w.Write(literal.LBRACK)
+		for i, j := range d.ListValues[value.Ref].Refs {
+			err = d.PrintValue(d.Value(j), w)
+			if err != nil {
+				return
+			}
+			if i != len(d.ListValues[value.Ref].Refs)-1 {
+				_, err = w.Write(literal.COMMA)
+			}
+		}
+		_, err = w.Write(literal.RBRACK)
+	case ValueKindObject:
+		_, err = w.Write(literal.LBRACE)
+		for i, j := range d.ObjectValues[value.Ref].Refs {
+			_, err = w.Write(d.ObjectFieldNameBytes(j))
+			if err != nil {
+				return
+			}
+			_, err = w.Write(literal.COLON)
+			if err != nil {
+				return
+			}
+			_, err = w.Write(literal.SPACE)
+			if err != nil {
+				return
+			}
+			err = d.PrintValue(d.ObjectFieldValue(j), w)
+			if err != nil {
+				return
+			}
+			if i != len(d.ObjectValues[value.Ref].Refs)-1 {
+				_, err = w.Write(literal.COMMA)
+				if err != nil {
+					return
+				}
+			}
+		}
+		_, err = w.Write(literal.RBRACE)
+	case ValueKindEnum:
+		_, err = w.Write(d.Input.ByteSlice(d.EnumValues[value.Ref].Name))
+	}
+	return
+}
+
+func (d *Document) PrintValueBytes(value Value, buf []byte) ([]byte, error) {
+	if buf == nil {
+		buf = make([]byte, 0, 24)
+	}
+	b := bytes.NewBuffer(buf)
+	err := d.PrintValue(value, b)
+	return b.Bytes(), err
+}
+
+func (d *Document) Value(ref int) Value {
+	return d.Values[ref]
+}
+
+func (d *Document) ValuesAreEqual(left, right Value) bool {
+	if left.Kind != right.Kind {
+		return false
+	}
+	switch left.Kind {
+	case ValueKindString:
+		return d.StringValuesAreEquals(left.Ref, right.Ref)
+	case ValueKindBoolean:
+		return d.BooleanValuesAreEqual(left.Ref, right.Ref)
+	case ValueKindInteger:
+		return d.IntValuesAreEquals(left.Ref, right.Ref)
+	case ValueKindFloat:
+		return d.FloatValuesAreEqual(left.Ref, right.Ref)
+	case ValueKindVariable:
+		return d.VariableValuesAreEqual(left.Ref, right.Ref)
+	case ValueKindNull:
+		return true
+	case ValueKindList:
+		return d.ListValuesAreEqual(left.Ref, right.Ref)
+	case ValueKindObject:
+		return d.ObjectValuesAreEqual(left.Ref, right.Ref)
+	case ValueKindEnum:
+		return d.EnumValuesAreEqual(left.Ref, right.Ref)
+	default:
+		return false
+	}
+}
